The Value firled for measurement to integer

+++ Aidn.Shared/Models/NewsScore.cs
@@ -1,4 +1,4 @@
 ﻿namespace Aidn.Shared.Models;
 
-public record NewsScore(double Score);
+public record NewsScore(int Score);
